Make threshold configurable for TreeOverlay
