+++ consensus-engine/src/overlay/tree_overlay/overlay.rs
@@ -1,6 +1,10 @@
 use super::tree::Tree;
+use crate::overlay::threshold::{
+    apply_threshold, default_leader_super_majority_threshold, deser_fraction,
+};
 use crate::overlay::CommitteeMembership;
 use crate::{overlay::LeaderSelection, Committee, NodeId, Overlay};
+use fraction::Fraction;
 use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Clone)]
@@ -11,6 +15,11 @@ pub struct TreeOverlaySettings<L: LeaderSelection, M: CommitteeMembership> {
     pub number_of_committees: usize,
     pub leader: L,
     pub committee_membership: M,
+    /// A fraction representing the threshold in the form `<num>/<den>'
+    /// Defaults to 2/3
+    #[serde(with = "deser_fraction")]
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub leader_super_majority_threshold: Option<Fraction>,
 }
 
 #[derive(Debug, Clone)]
@@ -21,6 +30,7 @@ pub struct TreeOverlay<L, M> {
     pub(super) carnot_tree: Tree,
     pub(super) leader: L,
     pub(super) committee_membership: M,
+    pub(super) leader_threshold: Fraction,
 }
 
 impl<L, M> Overlay for TreeOverlay<L, M>
@@ -40,6 +50,7 @@ where
             number_of_committees,
             leader,
             committee_membership,
+            leader_super_majority_threshold,
         } = settings;
 
         committee_membership.reshape_committees(&mut nodes);
@@ -52,6 +63,8 @@ where
             carnot_tree,
             leader,
             committee_membership,
+            leader_threshold: leader_super_majority_threshold
+                .unwrap_or_else(default_leader_super_majority_threshold),
         }
     }
 
@@ -158,8 +171,7 @@ where
         // });
         // let root_size = self.root_committee().len();
         // let committee_size = root_size + children_size;
-        let committee_size = self.root_committee().len();
-        (committee_size * 2 / 3) + 1
+        apply_threshold(self.root_committee().len(), self.leader_threshold)
     }
 
     fn update_leader_selection<F, E>(&self, f: F) -> Result<Self, E>
@@ -186,6 +198,7 @@ where
                 number_of_committees: self.number_of_committees,
                 leader: self.leader.clone(),
                 committee_membership,
+                leader_super_majority_threshold: Some(self.leader_threshold),
             };
             Self::new(settings)
         })
@@ -204,6 +217,7 @@ where
             number_of_committees: self.number_of_committees,
             leader,
             committee_membership,
+            leader_super_majority_threshold: Some(self.leader_threshold),
         })
     }
 
@@ -234,6 +248,7 @@ mod tests {
             number_of_committees: 3,
             leader: RoundRobin::new(),
             committee_membership: FisherYatesShuffle::new(ENTROPY),
+            leader_super_majority_threshold: None,
         });
 
         assert_eq!(*overlay.leader(), nodes[0]);
@@ -248,6 +263,7 @@ mod tests {
             number_of_committees: 3,
             leader: RoundRobin::new(),
             committee_membership: FisherYatesShuffle::new(ENTROPY),
+            leader_super_majority_threshold: None,
         });
 
         let leader = overlay.next_leader();
@@ -265,6 +281,7 @@ mod tests {
             number_of_committees: 3,
             leader: RoundRobin::new(),
             committee_membership: FisherYatesShuffle::new(ENTROPY),
+            leader_super_majority_threshold: None,
         });
 
         let mut expected_root = Committee::new();
@@ -283,6 +300,7 @@ mod tests {
             number_of_committees: 3,
             leader: RoundRobin::new(),
             committee_membership: FisherYatesShuffle::new(ENTROPY),
+            leader_super_majority_threshold: None,
         });
 
         let mut leaf_committees = overlay
@@ -313,6 +331,7 @@ mod tests {
             number_of_committees: 3,
             leader: RoundRobin::new(),
             committee_membership: FisherYatesShuffle::new(ENTROPY),
+            leader_super_majority_threshold: None,
         });
 
         assert_eq!(overlay.super_majority_threshold(overlay.nodes[8]), 0);
@@ -327,6 +346,7 @@ mod tests {
             number_of_committees: 3,
             leader: RoundRobin::new(),
             committee_membership: FisherYatesShuffle::new(ENTROPY),
+            leader_super_majority_threshold: None,
         });
 
         assert_eq!(overlay.super_majority_threshold(overlay.nodes[0]), 3);
@@ -341,6 +361,7 @@ mod tests {
             number_of_committees: 3,
             leader: RoundRobin::new(),
             committee_membership: FisherYatesShuffle::new(ENTROPY),
+            leader_super_majority_threshold: None,
         });
 
         assert_eq!(
